requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
mailchimp-marketing==3.0.12
Markdown==3.2.2
oauthlib==3.1.0
py-gfm==1.0.0
py-trello==0.17.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
